Update nelder_mead.d
It works finally for the dot product !
However, it works only when the simplex is fixed. This might means that we need another algorithm to generate a random simplex.
The better would be to generate a perfect simplex (distance between every point is the same) with a random orientation.
I didn't found any such algorithm, seems interesting though. Maybe it's equivalent to the uniform generation of a random rotation in R^n.